+++ Tests/OfflineTests/ClosureSynchronizerOfflineTests.swift
@@ -756,15 +756,6 @@ class ClosureSynchronizerOfflineTests: XCTestCase {
         wait(for: [expectation], timeout: 0.5)
     }
 
-    func testGetShieldedBalanceDeprecatedSucceed() {
-        synchronizerMock.getShieldedBalanceAccountIndexDeprecatedClosure = { receivedAccountIndex in
-            XCTAssertEqual(receivedAccountIndex, 3)
-            return 333
-        }
-
-        XCTAssertEqual(synchronizer.getShieldedBalance(accountIndex: 3), 333)
-    }
-
     func testGetShieldedBalanceSucceed() {
         synchronizerMock.getShieldedBalanceAccountIndexClosure = { receivedAccountIndex in
             XCTAssertEqual(receivedAccountIndex, 3)