+++ Tests/OfflineTests/CombineSynchronizerOfflineTests.swift
@@ -1017,15 +1017,6 @@ class CombineSynchronizerOfflineTests: XCTestCase {
         wait(for: [expectation], timeout: 0.5)
     }
 
-    func testGetShieldedBalanceDeprecatedSucceed() {
-        synchronizerMock.getShieldedBalanceAccountIndexDeprecatedClosure = { receivedAccountIndex in
-            XCTAssertEqual(receivedAccountIndex, 3)
-            return 333
-        }
-
-        XCTAssertEqual(synchronizer.getShieldedBalance(accountIndex: 3), 333)
-    }
-
     func testGetShieldedBalanceSucceed() {
         synchronizerMock.getShieldedBalanceAccountIndexClosure = { receivedAccountIndex in
             XCTAssertEqual(receivedAccountIndex, 3)