+++ Tests/TestUtils/SynchronizerMock.swift
@@ -152,11 +152,6 @@ class SynchronizerMock: Synchronizer {
         return try await getTransparentBalanceAccountIndexClosure(accountIndex)
     }
 
-    var getShieldedBalanceAccountIndexDeprecatedClosure: ((Int) -> Int64)! = nil
-    func getShieldedBalance(accountIndex: Int) -> Int64 {
-        return getShieldedBalanceAccountIndexDeprecatedClosure(accountIndex)
-    }
-
     var getShieldedBalanceAccountIndexClosure: ((Int) -> Zatoshi)! = nil
     func getShieldedBalance(accountIndex: Int) -> Zatoshi {
         return getShieldedBalanceAccountIndexClosure(accountIndex)